+++ cart_service/src/main/java/on/ssgdeal/cart_service/domain/repository/CartRepository.java
@@ -1,7 +1,7 @@
 package on.ssgdeal.cart_service.domain.repository;
 
 import java.util.Optional;
-import on.ssgdeal.cart_service.application.service.dto.UpdateCartProductDto;
+import on.ssgdeal.cart_service.infrastructure.persistence.repository.dto.UpdateCartProductDto;
 import on.ssgdeal.cart_service.domain.entity.CartProduct;
 import on.ssgdeal.cart_service.infrastructure.persistence.repository.dto.AddCartProductDto;
